SnapOnMove: moved TestAcq code to mmstudio.  Fix NullPointer Exception.
